pyproject.toml
[project]
name = "portal_multiespectral"
version = "0.1.0"
requires-python = ">=3.8,<3.10"
dependencies = [
    "fastapi",
    "python-multipart",
    "uvicorn",
    "fastapi-structlog",
    "asgi-correlation-id",
    "pillow",
    "bcrypt==3.2.0",
    "numpy<1.28",
    "tensorflow<2.11",
    "matplotlib",
    "scikit-learn",
    "rasterio",
    "shapely",
    "asyncpg",
    "passlib",
    "zxcvbn",
    "python-jose",
    "geopandas",
    "earthengine-api",
    "aiohttp",
    "aiofiles",
    "natsort",
    "geemap",
    "protobuf==3.20.*"
]

[tool.setuptools]
packages = ["database", "features", "middleware", "utils"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
